# src/main/java/org/siri_hate/chat_service/controller/ChatController.java
package org.siri_hate.chat_service.controller;

import org.siri_hate.chat_service.service.ChatService;
import org.siri_hate.chat_service.api.ChatApi;
import org.siri_hate.chat_service.dto.*;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
public class ChatController implements ChatApi {

    private final ChatService chatService;

    @Autowired
    public ChatController(ChatService chatService) {
        this.chatService = chatService;
    }

    @Override
    public ResponseEntity<ChatFullResponseDTO> createChat(String xUserName, ChatRequestDTO chatRequestDTO) {
        var response = chatService.createChat(xUserName, chatRequestDTO);
        return new ResponseEntity<>(response, HttpStatus.CREATED);
    }

    @Override
    public ResponseEntity<Void> deleteChat(Long id) {
        chatService.deleteChat(id);
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }

    @Override
    public ResponseEntity<ChatSummaryResponseDTO> getChatById(Long id) {
        var response = chatService.getChat(id);
        return new ResponseEntity<>(response, HttpStatus.OK);
    }

    @Override
    public ResponseEntity<ChatPageResponseDTO> getMyChats(Integer page, Integer size, String xUserName) {
        var response = chatService.getMyChats(xUserName, page, size);
        return new ResponseEntity<>(response, HttpStatus.OK);
    }

    @Override
    public ResponseEntity<List<Long>> getMutedChats(String xUserName) {
        return new ResponseEntity<>(chatService.getMutedChats(xUserName), HttpStatus.OK);
    }

    @Override
    public ResponseEntity<Void> muteChatNotifications(Long id, String xUserName) {
        chatService.muteChat(id, xUserName);
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }

    @Override
    public ResponseEntity<Void> unmuteChatNotifications(Long id, String xUserName) {
        chatService.unmuteChat(id, xUserName);
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }
}